# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "E-commerce backend API for household chemicals, cosmetics, and home goods.",
        "title": "Commerce Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/time-slots": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all time slots including inactive (admin)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Time Slots"
                ],
                "summary": "List all time slots (admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.TimeSlot"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new time slot (admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Time Slots"
                ],
                "summary": "Create time slot",
                "parameters": [
                    {
                        "description": "Time slot data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.CreateTimeSlotRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.TimeSlot"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/time-slots/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing time slot (admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Time Slots"
                ],
                "summary": "Update time slot",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Time slot ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Time slot data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.UpdateTimeSlotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.TimeSlot"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a time slot (admin)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Time Slots"
                ],
                "summary": "Delete time slot",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Time slot ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload an image file (admin)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Upload file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image file (jpg, jpeg, png, gif, webp, svg)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an uploaded file (admin)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Delete file",
                "parameters": [
                    {
                        "description": "JSON with url field",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/time-slots": {
            "get": {
                "description": "Get all active time slots (public)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Time Slots"
                ],
                "summary": "List time slots",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.TimeSlotResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_meransoft_commerce-backend_internal_models.ApiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_meransoft_commerce-backend_internal_models.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.ApiError"
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_meransoft_commerce-backend_internal_models.Pagination"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_meransoft_commerce-backend_internal_models.CreateTimeSlotRequest": {
            "type": "object",
            "required": [
                "end_time",
                "label",
                "start_time"
            ],
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "label": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
...
        "github_com_meransoft_commerce-backend_internal_models.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_meransoft_commerce-backend_internal_models.TimeSlot": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "label": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "github_com_meransoft_commerce-backend_internal_models.TimeSlotResponse": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
...
        "github_com_meransoft_commerce-backend_internal_models.UpdateTimeSlotRequest": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "label": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}